requirements.txt
altgraph==0.17.4
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
fonttools==4.60.1
idna==3.11
Js2Py==0.74
kiwisolver==1.4.9
matplotlib @ file:///C:/Users/Sathaniel/Desktop/PROGRAMACION/PROYECTOS/micro_remoto/matplotlib/matplotlib-3.10.7-cp313-cp313-win_amd64.whl#sha256=744991e0cc863dd669c8dc9136ca4e6e0082be2070b9d793cbd64bec872a6815
numpy @ file:///C:/Users/Sathaniel/Desktop/PROGRAMACION/numpy/numpy-2.3.3-cp313-cp313-win_amd64.whl#sha256=f0dadeb302887f07431910f67a14d57209ed91130be0adea2f9793f1a4f817cf
packaging==25.0
pefile==2023.2.7
pillow @ file:///C:/Users/Sathaniel/Desktop/PROGRAMACION/PROYECTOS/micro_remoto/pillow/pillow-11.3.0-cp313-cp313-win_amd64.whl#sha256=0bce5c4fd0921f99d2e858dc4d4d64193407e1b99478bc5cacecba2311abde51
pipwin @ file:///C:/Users/Sathaniel/Desktop/PROGRAMACION/PROYECTOS/micro_remoto/pipwin/pipwin-0.5.2.tar.gz#sha256=34791e633760240ac970ef41bf4cee4c31616ba31d9f58112e9bb83bab276746
PyAudio @ file:///C:/Users/Sathaniel/Desktop/PROGRAMACION/pyaudio/PyAudio-0.2.14-cp313-cp313-win_amd64.whl#sha256=692d8c1446f52ed2662120bcd9ddcb5aa2b71f38bda31e58b19fb4672fffba69
pyinstaller==6.16.0
pyinstaller-hooks-contrib @ file:///C:/Users/Sathaniel/Desktop/PROGRAMACION/PROYECTOS/micro_remoto/pyinstaller/pyinstaller_hooks_contrib-2025.9-py3-none-any.whl#sha256=ccbfaa49399ef6b18486a165810155e5a8d4c59b41f20dc5da81af7482aaf038
pyjsparser==2.7.1
pyparsing==3.2.5
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
